Reorganized files into proper folders
