+++ src/main/java/com/lambdaschool/wellness/service/Cloudinary/ImageController.java
@@ -1,35 +1,24 @@
-package com.lambdaschool.wellness.controllers;
+package com.lambdaschool.wellness.service.Cloudinary;
 
 import com.cloudinary.Cloudinary;
 import com.cloudinary.utils.ObjectUtils;
-import com.lambdaschool.wellness.model.Image;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
 import java.util.Map;
 
-
 @RestController
 @RequestMapping("/image")
 @CrossOrigin("http://localhost:3000")
 public class ImageController {
-    Cloudinary cloudinary = new Cloudinary(ObjectUtils.asMap(
-            "cloud_name", System.getenv("CLOUDINARY_CLOUD_NAME"),
-            "api_key", System.getenv("CLOUDINARY_API_KEY"),
-            "api_secret", System.getenv("CLOUDINARY_API_SECRET_KEY")
-    ));
-
+    Cloudinary cloudinary = new Cloudinary(ObjectUtils.asMap("cloud_name", System.getenv("CLOUDINARY_CLOUD_NAME"),
+            "api_key", System.getenv("CLOUDINARY_API_KEY"), "api_secret", System.getenv("CLOUDINARY_API_SECRET_KEY")));
 
     @PostMapping("/signed-upload")
     public Map uploadSignedImage(@RequestBody Image data) throws IOException {
-//        System.out.println("BASE64IMAGE DATA: "+ data.getBase64ImageData());
-//        System.out.println("FILE PARAMS DATA: "+ data.getFileParams().toString());
+        // System.out.println("BASE64IMAGE DATA: "+ data.getBase64ImageData());
+        // System.out.println("FILE PARAMS DATA: "+ data.getFileParams().toString());
         Map result = cloudinary.uploader().upload(data.getBase64ImageData(), data.getFileParams());
         return result;
-
     }
 }
-
-
-
-
